BZ-1187783 - kie-config-cli: remove-deployment & push-changes does not work - warn in case deployment was not found - remove-deployment command
(cherry picked from commit 15ffb96ea949d22fd1c65363a52328245127a147)

+++ kie-config-cli/src/main/java/org/kie/config/cli/command/impl/RemoveDeploymentConfigCliCommand.java
@@ -54,6 +54,10 @@ public class RemoveDeploymentConfigCliCommand implements CliCommand {
 		String ksession = input.nextLine();
 		
 		KModuleDeploymentUnit unit = new KModuleDeploymentUnit(groupId, artifactId, version, kbase, ksession);
+
+		if (deploymentConfigService.getDeployment(unit.getIdentifier()) == null) {
+			return "No deployment found with id " + unit.getIdentifier();
+		}
 		
 		deploymentConfigService.removeDeployment(unit.getIdentifier());
